Added user auth, and search filters, Publish dates, FB comments and likes

+++ posts/views.py
@@ -1,16 +1,21 @@
 from django.shortcuts import render, get_object_or_404, redirect
-from django.http import HttpResponse, HttpResponseRedirect
+from django.http import HttpResponse, HttpResponseRedirect, Http404
 from django.contrib import messages
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
+from django.utils import timezone
+from django.db.models import Q
 from .models import Post
 from .forms import PostForm
 
 
 
 def post_create(request):
-    form = PostForm(request.POST or None)
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
+    form = PostForm(request.POST or None, request.FILES or None)
     if form.is_valid():
         post = form.save(commit=False)
+        post.user = request.user
         post.save()
         messages.success(request, "Successfully Created!")
         return HttpResponseRedirect(post.get_absolute_url())
@@ -19,14 +24,27 @@ def post_create(request):
     return render(request, 'post_form.html', {'form': form})
 
 
-def post_detail(request, post_id):
-    post = get_object_or_404(Post, id=post_id)
+def post_detail(request, slug):
+    post = get_object_or_404(Post, slug=slug)
+    if post.draft or post.publish > timezone.now().date():
+        if not request.user.is_staff or not request.user.is_superuser:
+            raise Http404
     return render(request, 'post_detail.html', {'post': post})
 
 
 def post_list(request):
-    all_posts = Post.objects.all()
-    paginator = Paginator(all_posts, 4)
+    today = timezone.now().date()
+    all_posts = Post.objects.active()
+    if request.user.is_staff or request.user.is_superuser:
+        all_posts = Post.objects.all()
+    query = request.GET.get("q")
+    if query:
+        all_posts = all_posts.filter(
+            Q(title__icontains=query) |
+            Q(content__icontains=query) |
+            Q(user__username__icontains=query)
+        ).distinct()
+    paginator = Paginator(all_posts, 1)
     page_var = "p"
     page = request.GET.get(page_var)
 
@@ -37,12 +55,14 @@ def post_list(request):
     except EmptyPage:
         posts = paginator.page(paginator.num_pages)
 
-    return render(request, 'post_list.html', {'posts': posts, 'page_var': page_var})
+    return render(request, 'post_list.html', {'posts': posts, 'page_var': page_var, 'today': today})
 
 
-def post_update(request, post_id):
-    post = get_object_or_404(Post, id=post_id)
-    form = PostForm(request.POST or None, instance=post)
+def post_update(request, slug):
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
+    post = get_object_or_404(Post, slug=slug)
+    form = PostForm(request.POST or None, request.FILES or None, instance=post)
     if form.is_valid():
         post = form.save(commit=False)
         post.save()
@@ -51,8 +71,10 @@ def post_update(request, post_id):
     return render(request, 'post_form.html', {'post': post, 'form': form})
 
 
-def post_delete(request, post_id):
-    post = get_object_or_404(Post, id=post_id)
+def post_delete(request, slug):
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
+    post = get_object_or_404(Post, slug=slug)
     post.delete()
     messages.success(request, "Successfully Deleted!")
     return redirect("posts:list")
